Cleanup : code cleanup and comments in the homefragment and RecyclerViewLocalMaps
